Fix Tags & Unique Client Feedback
Tags are dependent on a data point to place them at so a little
maneuvering is necessary to be sure the appropriate data point is in
the right place in the timeline.

If a student gives repeated feedback, it should replace the last
feedback given by that student, not accumulate as if different students
are giving feedback.